优化

+++ TY/TYModel/Ren.cs
@@ -7,9 +7,9 @@ namespace TYModel
 {
     public class Ren
     {
-        public int id { get; set; }
+        public long id { get; set; }
         public string name { get; set; }
-        public int age { get; set; }
+        public long age { get; set; }
         public string sfzhm { get; set; }
     }
 }
